REF/TST: collect index tests
