+++ pandas/tests/indexes/test_index_new.py
@@ -272,3 +272,14 @@ class TestIndexConstructorUnwrapping:
         expected = Index(array)
         result = Index(ArrayLike(array))
         tm.assert_index_equal(result, expected)
+
+
+class TestIndexConstructionErrors:
+    def test_constructor_overflow_int64(self):
+        # see GH#15832
+        msg = (
+            "The elements provided in the data cannot "
+            "all be casted to the dtype int64"
+        )
+        with pytest.raises(OverflowError, match=msg):
+            Index([np.iinfo(np.uint64).max - 1], dtype="int64")
